+++ front/src/pages/trade/chart/candle/CandleGraph.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import userAtom, { IUser } from '@recoil/user';
-import { IChartItem } from '@recoil/chart';
+import { IChartItem } from '@src/types';
 
 import { IGraphComponentProps, TTheme, CANDLE_GAP, getMaxValue, getMinValue, getPriceColor } from '../common';
 import CandleBackground from './CandleBackground';
