+++ front/src/pages/trade/chart/hooks/useChart.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect } from 'react';
 import { useRecoilState } from 'recoil';
+import { IChartItem } from '@src/types';
+import chartAtom from '@recoil/chart';
 import { IStockListItem, IStockChartItem } from '@recoil/stockList/index';
-import chartAtom, { IChartItem } from '@recoil/chart/index';
 import { TChartType, MAX_NUM_OF_CANDLES } from '../common';
 import fetchChartData from '../fetchChartData';
