+++ front/src/pages/trade/chart/volume/VolumeLegend.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import userAtom, { IUser } from '@recoil/user';
-import { IChartItem } from '@recoil/chart';
+import { IChartItem } from '@src/types';
 import {
 	ICrossLine,
 	TTheme,
