+++ front/src/recoil/chart/index.ts
@@ -1,5 +1,1 @@
-import atom from './atom';
-import type { IChartItem } from './atom';
-
-export { IChartItem };
-export default atom;
+export { default } from './atom';
